chore: split ProductGallery logic into section and ui component
