+++ packages/core/src/components/ui/ProductGallery/ProductGalleryPage.tsx
