+++ packages/core/src/components/ui/ProductGallery/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ProductGallery'
