+++ packages/core/src/components/ui/ProductGallery/useDelayedFacets.ts
