+++ packages/core/src/components/ui/ProductGallery/usePageProducts.ts
